+++ R/init.R
@@ -37,6 +37,7 @@ init_site <- function(pkg = ".") {
   rule("Initialising site")
   dir_create(pkg$dst_path)
   copy_assets(pkg)
+  build_favicons(pkg)
   copy_favicons(pkg)
 
   build_site_meta(pkg)