+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgdown 1.3.0.9000 (development version)
 
+* Favicons are now created by `build_site()` from a package logo. Run `build_favicons(clobber = TRUE)` to re-create favicons if the logo changes.
+
 * Function names can now be included in headers without spurious auto-linking (#948).
 
 * The title and description of the homepage now corresponds by default to an unquoted version of the DESCRIPTION Title and Description. Furthermore, one can override the title and description of the homepage via the `title` and `description` fields in the home section of config (#957, @maelle).